ui: Move wildcard param decoding to routlet service
