+++ ui/packages/consul-ui/app/routing/route.js
@@ -2,17 +2,14 @@ import Route from '@ember/routing/route';
 import { get, setProperties, action } from '@ember/object';
 import { inject as service } from '@ember/service';
 
-// paramsFor
 import { routes } from 'consul-ui/router';
-import wildcard from 'consul-ui/utils/routing/wildcard';
-
-const isWildcard = wildcard(routes);
 
 export default class BaseRoute extends Route {
   @service('container') container;
   @service('env') env;
   @service('repository/permission') permissions;
   @service('router') router;
+  @service('routlet') routlet;
 
   _setRouteName() {
     super._setRouteName(...arguments);
@@ -114,27 +111,14 @@ export default class BaseRoute extends Route {
   }
 
   /**
-   * Adds urldecoding to any wildcard route `params` passed into ember `model`
-   * hooks, plus of course anywhere else where `paramsFor` is used. This means
-   * the entire ember app is now changed so that all paramsFor calls returns
-   * urldecoded params instead of raw ones.
-   * For example we use this largely for URLs for the KV store:
-   * /kv/*key > /ui/kv/%25-kv-name/%25-here > key = '%-kv-name/%-here'
+   * Normalizes any params passed into ember `model` hooks, plus of course
+   * anywhere else where `paramsFor` is used. This means the entire ember app
+   * is now changed so that all paramsFor calls returns normalized params
+   * instead of raw ones. For example we use this largely for URLs for the KV
+   * store: /kv/*key > /ui/kv/%25-kv-name/%25-here > key = '%-kv-name/%-here'
    */
   paramsFor(name) {
-    const params = super.paramsFor(...arguments);
-    if (isWildcard(this.routeName)) {
-      return Object.keys(params).reduce(function(prev, item) {
-        if (typeof params[item] !== 'undefined') {
-          prev[item] = decodeURIComponent(params[item]);
-        } else {
-          prev[item] = params[item];
-        }
-        return prev;
-      }, {});
-    } else {
-      return params;
-    }
+    return this.routlet.normalizeParamsFor(this.routeName, super.paramsFor(...arguments));
   }
 
   @action
